@@ -10,5 +10,5 @@ class DragAndDropItem implements DragAndDropInterface {
   /// Set to false if it must remain fixed.
   final bool canDrag;
 
-  DragAndDropItem({@required this.child, this.canDrag = true});
+  DragAndDropItem({required this.child, this.canDrag = true});
 }
