@@ -6,16 +6,16 @@ import 'package:flutter/widgets.dart';
 
 class DragAndDropItemTarget extends StatefulWidget {
   final Widget child;
-  final DragAndDropListInterface parent;
-  final DragAndDropBuilderParameters parameters;
-  final OnItemDropOnLastTarget onReorderOrAdd;
+  final DragAndDropListInterface? parent;
+  final DragAndDropBuilderParameters? parameters;
+  final OnItemDropOnLastTarget? onReorderOrAdd;
 
   DragAndDropItemTarget(
-      {@required this.child,
-      @required this.onReorderOrAdd,
-      @required this.parameters,
+      {required this.child,
+      required this.onReorderOrAdd,
+      required this.parameters,
       this.parent,
-      Key key})
+      Key? key})
       : super(key: key);
 
   @override
@@ -24,32 +24,29 @@ class DragAndDropItemTarget extends StatefulWidget {
 
 class _DragAndDropItemTarget extends State<DragAndDropItemTarget>
     with TickerProviderStateMixin {
-  DragAndDropItem _hoveredDraggable;
+  DragAndDropItem? _hoveredDraggable;
 
   @override
   Widget build(BuildContext context) {
     return Stack(
       children: <Widget>[
         Column(
-          crossAxisAlignment: widget.parameters.verticalAlignment,
+          crossAxisAlignment: widget.parameters!.verticalAlignment,
           children: <Widget>[
             AnimatedSize(
               duration: Duration(
-                  milliseconds: widget.parameters.itemSizeAnimationDuration),
+                  milliseconds: widget.parameters!.itemSizeAnimationDuration),
               vsync: this,
               alignment: Alignment.bottomCenter,
               child: _hoveredDraggable != null
                   ? Opacity(
-                      opacity: widget.parameters.itemGhostOpacity,
-                      child: widget.parameters.itemGhost ??
-                          _hoveredDraggable.child,
+                      opacity: widget.parameters!.itemGhostOpacity,
+                      child: widget.parameters!.itemGhost ??
+                          _hoveredDraggable!.child,
                     )
                   : Container(),
             ),
-            widget.child ??
-                Container(
-                  height: 20,
-                ),
+            widget.child
           ],
         ),
         Positioned.fill(
@@ -60,9 +57,9 @@ class _DragAndDropItemTarget extends State<DragAndDropItemTarget>
             },
             onWillAccept: (incoming) {
               bool accept = true;
-              if (widget.parameters.itemTargetOnWillAccept != null)
+              if (widget.parameters!.itemTargetOnWillAccept != null)
                 accept =
-                    widget.parameters.itemTargetOnWillAccept(incoming, widget);
+                    widget.parameters!.itemTargetOnWillAccept!(incoming, widget);
               if (accept && mounted) {
                 setState(() {
                   _hoveredDraggable = incoming;
@@ -81,7 +78,7 @@ class _DragAndDropItemTarget extends State<DragAndDropItemTarget>
               if (mounted) {
                 setState(() {
                   if (widget.onReorderOrAdd != null)
-                    widget.onReorderOrAdd(incoming, widget.parent, widget);
+                    widget.onReorderOrAdd!(incoming, widget.parent, widget);
                   _hoveredDraggable = null;
                 });
               }
