@@ -8,10 +8,10 @@ import 'package:flutter/widgets.dart';
 
 class DragAndDropItemWrapper extends StatefulWidget {
   final DragAndDropItem child;
-  final DragAndDropBuilderParameters parameters;
+  final DragAndDropBuilderParameters? parameters;
 
   DragAndDropItemWrapper(
-      {@required this.child, @required this.parameters, Key key})
+      {required this.child, required this.parameters, Key? key})
       : super(key: key);
 
   @override
@@ -20,7 +20,7 @@ class DragAndDropItemWrapper extends StatefulWidget {
 
 class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
     with TickerProviderStateMixin {
-  DragAndDropItem _hoveredDraggable;
+  DragAndDropItem? _hoveredDraggable;
 
   bool _dragging = false;
   Size _containerSize = Size.zero;
@@ -30,37 +30,37 @@ class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
   Widget build(BuildContext context) {
     Widget draggable;
     if (widget.child.canDrag) {
-      if (widget.parameters.dragHandle != null) {
+      if (widget.parameters!.dragHandle != null) {
         Widget feedback = Container(
-          width: widget.parameters.itemDraggingWidth ?? _containerSize.width,
+          width: widget.parameters!.itemDraggingWidth ?? _containerSize.width,
           child: Stack(
             children: [
               widget.child.child,
               Positioned(
-                right: widget.parameters.dragHandleOnLeft ? null : 0,
-                left: widget.parameters.dragHandleOnLeft ? 0 : null,
-                top: widget.parameters.itemDragHandleVerticalAlignment ==
+                right: widget.parameters!.dragHandleOnLeft ? null : 0,
+                left: widget.parameters!.dragHandleOnLeft ? 0 : null,
+                top: widget.parameters!.itemDragHandleVerticalAlignment ==
                         DragHandleVerticalAlignment.bottom
                     ? null
                     : 0,
-                bottom: widget.parameters.itemDragHandleVerticalAlignment ==
+                bottom: widget.parameters!.itemDragHandleVerticalAlignment ==
                         DragHandleVerticalAlignment.top
                     ? null
                     : 0,
-                child: widget.parameters.dragHandle,
+                child: widget.parameters!.dragHandle!,
               ),
             ],
           ),
         );
 
         var positionedDragHandle = Positioned(
-          right: widget.parameters.dragHandleOnLeft ? null : 0,
-          left: widget.parameters.dragHandleOnLeft ? 0 : null,
-          top: widget.parameters.itemDragHandleVerticalAlignment ==
+          right: widget.parameters!.dragHandleOnLeft ? null : 0,
+          left: widget.parameters!.dragHandleOnLeft ? 0 : null,
+          top: widget.parameters!.itemDragHandleVerticalAlignment ==
                   DragHandleVerticalAlignment.bottom
               ? null
               : 0,
-          bottom: widget.parameters.itemDragHandleVerticalAlignment ==
+          bottom: widget.parameters!.itemDragHandleVerticalAlignment ==
                   DragHandleVerticalAlignment.top
               ? null
               : 0,
@@ -68,24 +68,24 @@ class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
             cursor: SystemMouseCursors.grab,
             child: Draggable<DragAndDropItem>(
               data: widget.child,
-              axis: widget.parameters.axis == Axis.vertical &&
-                      widget.parameters.constrainDraggingAxis
+              axis: widget.parameters!.axis == Axis.vertical &&
+                      widget.parameters!.constrainDraggingAxis
                   ? Axis.vertical
                   : null,
               child: MeasureSize(
                 onSizeChange: (size) {
                   setState(() {
-                    _dragHandleSize = size;
+                    _dragHandleSize = size!;
                   });
                 },
-                child: widget.parameters.dragHandle,
+                child: widget.parameters!.dragHandle,
               ),
               feedback: Transform.translate(
                 offset: _feedbackContainerOffset(),
                 child: Material(
                   color: Colors.transparent,
                   child: Container(
-                    decoration: widget.parameters.itemDecorationWhileDragging,
+                    decoration: widget.parameters!.itemDecorationWhileDragging,
                     child: feedback,
                   ),
                 ),
@@ -112,23 +112,23 @@ class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
             ],
           ),
         );
-      } else if (widget.parameters.dragOnLongPress) {
+      } else if (widget.parameters!.dragOnLongPress) {
         draggable = MeasureSize(
           onSizeChange: _setContainerSize,
           child: LongPressDraggable<DragAndDropItem>(
             data: widget.child,
-            axis: widget.parameters.axis == Axis.vertical &&
-                    widget.parameters.constrainDraggingAxis
+            axis: widget.parameters!.axis == Axis.vertical &&
+                    widget.parameters!.constrainDraggingAxis
                 ? Axis.vertical
                 : null,
             child: widget.child.child,
             feedback: Container(
               width:
-                  widget.parameters.itemDraggingWidth ?? _containerSize.width,
+                  widget.parameters!.itemDraggingWidth ?? _containerSize.width,
               child: Material(
                 child: Container(
                   child: widget.child.child,
-                  decoration: widget.parameters.itemDecorationWhileDragging,
+                  decoration: widget.parameters!.itemDecorationWhileDragging,
                 ),
                 color: Colors.transparent,
               ),
@@ -145,18 +145,18 @@ class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
           onSizeChange: _setContainerSize,
           child: Draggable<DragAndDropItem>(
             data: widget.child,
-            axis: widget.parameters.axis == Axis.vertical &&
-                    widget.parameters.constrainDraggingAxis
+            axis: widget.parameters!.axis == Axis.vertical &&
+                    widget.parameters!.constrainDraggingAxis
                 ? Axis.vertical
                 : null,
             child: widget.child.child,
             feedback: Container(
               width:
-                  widget.parameters.itemDraggingWidth ?? _containerSize.width,
+                  widget.parameters!.itemDraggingWidth ?? _containerSize.width,
               child: Material(
                 child: Container(
                   child: widget.child.child,
-                  decoration: widget.parameters.itemDecorationWhileDragging,
+                  decoration: widget.parameters!.itemDecorationWhileDragging,
                 ),
                 color: Colors.transparent,
               ),
@@ -172,7 +172,7 @@ class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
     } else {
       draggable = AnimatedSize(
         duration:
-            Duration(milliseconds: widget.parameters.itemSizeAnimationDuration),
+            Duration(milliseconds: widget.parameters!.itemSizeAnimationDuration),
         vsync: this,
         alignment: Alignment.bottomCenter,
         child: _hoveredDraggable != null ? Container() : widget.child.child,
@@ -182,26 +182,26 @@ class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
       children: <Widget>[
         Column(
           mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: widget.parameters.verticalAlignment,
+          crossAxisAlignment: widget.parameters!.verticalAlignment,
           children: <Widget>[
             AnimatedSize(
               duration: Duration(
-                  milliseconds: widget.parameters.itemSizeAnimationDuration),
+                  milliseconds: widget.parameters!.itemSizeAnimationDuration),
               vsync: this,
               alignment: Alignment.topLeft,
               child: _hoveredDraggable != null
                   ? Opacity(
-                      opacity: widget.parameters.itemGhostOpacity,
-                      child: widget.parameters.itemGhost ??
-                          _hoveredDraggable.child,
+                      opacity: widget.parameters!.itemGhostOpacity,
+                      child: widget.parameters!.itemGhost ??
+                          _hoveredDraggable!.child,
                     )
                   : Container(),
             ),
             Listener(
               child: draggable,
               onPointerMove: _onPointerMove,
-              onPointerDown: widget.parameters.onPointerDown,
-              onPointerUp: widget.parameters.onPointerUp,
+              onPointerDown: widget.parameters!.onPointerDown,
+              onPointerUp: widget.parameters!.onPointerUp,
             ),
           ],
         ),
@@ -213,9 +213,9 @@ class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
             },
             onWillAccept: (incoming) {
               bool accept = true;
-              if (widget.parameters.itemOnWillAccept != null)
+              if (widget.parameters!.itemOnWillAccept != null)
                 accept =
-                    widget.parameters.itemOnWillAccept(incoming, widget.child);
+                    widget.parameters!.itemOnWillAccept!(incoming, widget.child);
               if (accept && mounted) {
                 setState(() {
                   _hoveredDraggable = incoming;
@@ -233,8 +233,8 @@ class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
             onAccept: (incoming) {
               if (mounted) {
                 setState(() {
-                  if (widget.parameters.onItemReordered != null)
-                    widget.parameters.onItemReordered(incoming, widget.child);
+                  if (widget.parameters!.onItemReordered != null)
+                    widget.parameters!.onItemReordered!(incoming, widget.child);
                   _hoveredDraggable = null;
                 });
               }
@@ -248,12 +248,12 @@ class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
   Offset _feedbackContainerOffset() {
     double xOffset;
     double yOffset;
-    if (widget.parameters.dragHandleOnLeft) {
+    if (widget.parameters!.dragHandleOnLeft) {
       xOffset = 0;
     } else {
       xOffset = -_containerSize.width + _dragHandleSize.width;
     }
-    if (widget.parameters.itemDragHandleVerticalAlignment ==
+    if (widget.parameters!.itemDragHandleVerticalAlignment ==
         DragHandleVerticalAlignment.bottom) {
       yOffset = -_containerSize.height + _dragHandleSize.width;
     } else {
@@ -263,10 +263,10 @@ class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
     return Offset(xOffset, yOffset);
   }
 
-  void _setContainerSize(Size size) {
+  void _setContainerSize(Size? size) {
     if (mounted) {
       setState(() {
-        _containerSize = size;
+        _containerSize = size!;
       });
     }
   }
@@ -276,13 +276,13 @@ class _DragAndDropItemWrapper extends State<DragAndDropItemWrapper>
       setState(() {
         _dragging = dragging;
       });
-      if (widget.parameters.onItemDraggingChanged != null) {
-        widget.parameters.onItemDraggingChanged(widget.child, dragging);
+      if (widget.parameters!.onItemDraggingChanged != null) {
+        widget.parameters!.onItemDraggingChanged!(widget.child, dragging);
       }
     }
   }
 
   void _onPointerMove(PointerMoveEvent event) {
-    if (_dragging) widget.parameters.onPointerMove(event);
+    if (_dragging) widget.parameters!.onPointerMove!(event);
   }
 }
