@@ -8,28 +8,28 @@ import 'package:flutter/widgets.dart';
 
 class DragAndDropList implements DragAndDropListInterface {
   /// The widget that is displayed at the top of the list.
-  final Widget header;
+  final Widget? header;
 
   /// The widget that is displayed at the bottom of the list.
-  final Widget footer;
+  final Widget? footer;
 
   /// The widget that is displayed to the left of the list.
-  final Widget leftSide;
+  final Widget? leftSide;
 
   /// The widget that is displayed to the right of the list.
-  final Widget rightSide;
+  final Widget? rightSide;
 
   /// The widget to be displayed when a list is empty.
   /// If this is not null, it will override that set in [DragAndDropLists.contentsWhenEmpty].
-  final Widget contentsWhenEmpty;
+  final Widget? contentsWhenEmpty;
 
   /// The widget to be displayed as the last element in the list that will accept
   /// a dragged item.
-  final Widget lastTarget;
+  final Widget? lastTarget;
 
   /// The decoration displayed around a list.
   /// If this is not null, it will override that set in [DragAndDropLists.listDecoration].
-  final Decoration decoration;
+  final Decoration? decoration;
 
   /// The vertical alignment of the contents in this list.
   /// If this is not null, it will override that set in [DragAndDropLists.verticalAlignment].
@@ -41,7 +41,7 @@ class DragAndDropList implements DragAndDropListInterface {
 
   /// The child elements that will be contained in this list.
   /// It is possible to not provide any children when an empty list is desired.
-  final List<DragAndDropItem> children = List<DragAndDropItem>();
+  final List<DragAndDropItem> children = <DragAndDropItem>[];
 
   /// Whether or not this item can be dragged.
   /// Set to true if it can be reordered.
@@ -49,7 +49,7 @@ class DragAndDropList implements DragAndDropListInterface {
   final bool canDrag;
 
   DragAndDropList(
-      {List<DragAndDropItem> children,
+      {List<DragAndDropItem>? children,
       this.header,
       this.footer,
       this.leftSide,
@@ -66,20 +66,20 @@ class DragAndDropList implements DragAndDropListInterface {
   }
 
   @override
-  Widget generateWidget(DragAndDropBuilderParameters params) {
-    var contents = List<Widget>();
+  Widget generateWidget(DragAndDropBuilderParameters? params) {
+    var contents = <Widget>[];
     if (header != null) {
-      contents.add(Flexible(child: header));
+      contents.add(Flexible(child: header!));
     }
     Widget intrinsicHeight = IntrinsicHeight(
       child: Row(
         mainAxisAlignment: horizontalAlignment,
         mainAxisSize: MainAxisSize.max,
         crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: _generateDragAndDropListInnerContents(params),
+        children: _generateDragAndDropListInnerContents(params) as List<Widget>,
       ),
     );
-    if (params.axis == Axis.horizontal) {
+    if (params!.axis == Axis.horizontal) {
       intrinsicHeight = Container(
         width: params.listWidth,
         child: intrinsicHeight,
@@ -94,13 +94,13 @@ class DragAndDropList implements DragAndDropListInterface {
     contents.add(intrinsicHeight);
 
     if (footer != null) {
-      contents.add(Flexible(child: footer));
+      contents.add(Flexible(child: footer!));
     }
 
     return Container(
       width: params.axis == Axis.vertical
           ? double.infinity
-          : params.listWidth - params.listPadding.horizontal,
+          : params.listWidth - params.listPadding!.horizontal,
       decoration: decoration ?? params.listDecoration,
       child: Column(
         mainAxisSize: MainAxisSize.min,
@@ -110,15 +110,15 @@ class DragAndDropList implements DragAndDropListInterface {
     );
   }
 
-  List<Widget> _generateDragAndDropListInnerContents(
-      DragAndDropBuilderParameters params) {
-    var contents = List<Widget>();
+  List<Widget?> _generateDragAndDropListInnerContents(
+      DragAndDropBuilderParameters? params) {
+    var contents = <Widget?>[];
     if (leftSide != null) {
       contents.add(leftSide);
     }
     if (children != null && children.isNotEmpty) {
-      List<Widget> allChildren = List<Widget>();
-      if (params.addLastItemTargetHeightToTop) {
+      List<Widget?> allChildren = <Widget?>[];
+      if (params!.addLastItemTargetHeightToTop) {
         allChildren.add(Padding(
           padding: EdgeInsets.only(top: params.lastItemTargetHeight),
         ));
@@ -148,7 +148,7 @@ class DragAndDropList implements DragAndDropListInterface {
             child: Column(
               crossAxisAlignment: verticalAlignment,
               mainAxisSize: MainAxisSize.max,
-              children: allChildren,
+              children: allChildren as List<Widget>,
             ),
           ),
         ),
@@ -171,7 +171,7 @@ class DragAndDropList implements DragAndDropListInterface {
                 DragAndDropItemTarget(
                   parent: this,
                   parameters: params,
-                  onReorderOrAdd: params.onItemDropOnLastTarget,
+                  onReorderOrAdd: params!.onItemDropOnLastTarget,
                   child: lastTarget ??
                       Container(
                         height: params.lastItemTargetHeight,
