@@ -9,14 +9,13 @@ import 'package:drag_and_drop_lists/programmatic_expansion_tile.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-typedef void OnExpansionChanged(bool expanded);
 
 /// This class mirrors flutter's [ExpansionTile], with similar options.
 class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
-  final Widget title;
-  final Widget subtitle;
-  final Widget trailing;
-  final Widget leading;
+  final Widget? title;
+  final Widget? subtitle;
+  final Widget? trailing;
+  final Widget? leading;
   final bool initiallyExpanded;
 
   /// Set this to a unique key that will remain unchanged over the lifetime of the list.
@@ -24,11 +23,11 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
   final Key listKey;
 
   /// This function will be called when the expansion of a tile is changed.
-  final OnExpansionChanged onExpansionChanged;
-  final Color backgroundColor;
-  final List<DragAndDropItem> children;
-  final Widget contentsWhenEmpty;
-  final Widget lastTarget;
+  final OnExpansionChanged? onExpansionChanged;
+  final Color? backgroundColor;
+  final List<DragAndDropItem>? children;
+  final Widget? contentsWhenEmpty;
+  final Widget? lastTarget;
 
   /// Whether or not this item can be dragged.
   /// Set to true if it can be reordered.
@@ -53,7 +52,7 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
     this.onExpansionChanged,
     this.contentsWhenEmpty,
     this.lastTarget,
-    this.listKey,
+    required this.listKey,
     this.canDrag = true,
     this.disableTopAndBottomBorders = false,
   }) : assert(listKey != null) {
@@ -61,7 +60,7 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
   }
 
   @override
-  Widget generateWidget(DragAndDropBuilderParameters params) {
+  Widget generateWidget(DragAndDropBuilderParameters? params) {
     var contents = _generateDragAndDropListInnerContents(params);
 
     Widget expandable = ProgrammaticExpansionTile(
@@ -78,7 +77,7 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
       children: contents,
     );
 
-    if (params.listDecoration != null) {
+    if (params!.listDecoration != null) {
       expandable = Container(
         decoration: params.listDecoration,
         child: expandable,
@@ -87,7 +86,7 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
 
     if (params.listPadding != null) {
       expandable = Padding(
-        padding: params.listPadding,
+        padding: params.listPadding!,
         child: expandable,
       );
     }
@@ -95,11 +94,11 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
     Widget toReturn = ValueListenableBuilder(
       valueListenable: _expanded,
       child: expandable,
-      builder: (context, error, child) {
+      builder: (context, dynamic error, child) {
         if (!_expanded.value) {
           return Stack(
             children: <Widget>[
-              child,
+              child!,
               Positioned.fill(
                 child: DragTarget<DragAndDropItem>(
                   builder: (context, candidateData, rejectedData) {
@@ -119,7 +118,7 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
             ],
           );
         } else {
-          return child;
+          return child!;
         }
       },
     );
@@ -127,23 +126,23 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
     return toReturn;
   }
 
-  List<Widget> _generateDragAndDropListInnerContents(
-      DragAndDropBuilderParameters params) {
-    var contents = List<Widget>();
-    if (children != null && children.isNotEmpty) {
-      for (int i = 0; i < children.length; i++) {
+  List<Widget?> _generateDragAndDropListInnerContents(
+      DragAndDropBuilderParameters? params) {
+    var contents = <Widget>[];
+    if (children != null && children!.isNotEmpty) {
+      for (int i = 0; i < children!.length; i++) {
         contents.add(DragAndDropItemWrapper(
-          child: children[i],
+          child: children![i],
           parameters: params,
         ));
-        if (params.itemDivider != null && i < children.length - 1) {
-          contents.add(params.itemDivider);
+        if (params!.itemDivider != null && i < children!.length - 1) {
+          contents.add(params.itemDivider!);
         }
       }
       contents.add(DragAndDropItemTarget(
         parent: this,
         parameters: params,
-        onReorderOrAdd: params.onItemDropOnLastTarget,
+        onReorderOrAdd: params!.onItemDropOnLastTarget,
         child: lastTarget ??
             Container(
               height: params.lastItemTargetHeight,
@@ -163,7 +162,7 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
         DragAndDropItemTarget(
           parent: this,
           parameters: params,
-          onReorderOrAdd: params.onItemDropOnLastTarget,
+          onReorderOrAdd: params!.onItemDropOnLastTarget,
           child: lastTarget ??
               Container(
                 height: params.lastItemTargetHeight,
@@ -186,7 +185,7 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
   collapse() {
     if (!isExpanded) {
       _expanded.value = false;
-      _expansionKey.currentState.collapse();
+      _expansionKey.currentState!.collapse();
     }
   }
 
@@ -194,20 +193,20 @@ class DragAndDropListExpansion implements DragAndDropListExpansionInterface {
   expand() {
     if (!isExpanded) {
       _expanded.value = true;
-      _expansionKey.currentState.expand();
+      _expansionKey.currentState!.expand();
     }
   }
 
   _onSetExpansion(bool expanded) {
     _expanded.value = expanded;
 
-    if (onExpansionChanged != null) onExpansionChanged(expanded);
+    if (onExpansionChanged != null) onExpansionChanged!(expanded);
   }
 
   @override
   get isExpanded => _expanded.value;
 
-  Timer _expansionTimer;
+  late Timer _expansionTimer;
 
   _startExpansionTimer() async {
     _expansionTimer = Timer(Duration(milliseconds: 400), _expansionCallback);
