@@ -47,19 +47,19 @@ typedef void OnItemAdd(
 typedef void OnListAdd(DragAndDropListInterface newList, int newListIndex);
 typedef void OnListReorder(int oldListIndex, int newListIndex);
 typedef void OnListDraggingChanged(
-  DragAndDropListInterface list,
+  DragAndDropListInterface? list,
   bool dragging,
 );
 typedef bool ListOnWillAccept(
-  DragAndDropListInterface incoming,
-  DragAndDropListInterface target,
+  DragAndDropListInterface? incoming,
+  DragAndDropListInterface? target,
 );
 typedef void ListOnAccept(
   DragAndDropListInterface incoming,
   DragAndDropListInterface target,
 );
 typedef bool ListTargetOnWillAccept(
-    DragAndDropListInterface incoming, DragAndDropListTarget target);
+    DragAndDropListInterface? incoming, DragAndDropListTarget target);
 typedef void ListTargetOnAccept(
     DragAndDropListInterface incoming, DragAndDropListTarget target);
 typedef void OnItemDraggingChanged(
@@ -67,7 +67,7 @@ typedef void OnItemDraggingChanged(
   bool dragging,
 );
 typedef bool ItemOnWillAccept(
-  DragAndDropItem incoming,
+  DragAndDropItem? incoming,
   DragAndDropItem target,
 );
 typedef void ItemOnAccept(
@@ -75,7 +75,7 @@ typedef void ItemOnAccept(
   DragAndDropItem target,
 );
 typedef bool ItemTargetOnWillAccept(
-    DragAndDropItem incoming, DragAndDropItemTarget target);
+    DragAndDropItem? incoming, DragAndDropItemTarget target);
 typedef void ItemTargetOnAccept(
   DragAndDropItem incoming,
   DragAndDropListInterface parentList,
@@ -86,80 +86,80 @@ class DragAndDropLists extends StatefulWidget {
   /// The child lists to be displayed.
   /// If any of these children are [DragAndDropListExpansion] or inherit from
   /// [DragAndDropListExpansionInterface], [listGhost] must not be null.
-  final List<DragAndDropListInterface> children;
+  final List<DragAndDropListInterface>? children;
 
   /// Calls this function when a list element is reordered.
   /// Takes into account the index change when removing an item, so the
   /// [newItemIndex] can be used directly when inserting.
-  final OnItemReorder onItemReorder;
+  final OnItemReorder? onItemReorder;
 
   /// Calls this function when a list is reordered.
   /// Takes into account the index change when removing a list, so the
   /// [newListIndex] can be used directly when inserting.
-  final OnListReorder onListReorder;
+  final OnListReorder? onListReorder;
 
   /// Calls this function when a new item has been added.
-  final OnItemAdd onItemAdd;
+  final OnItemAdd? onItemAdd;
 
   /// Calls this function when a new list has been added.
-  final OnListAdd onListAdd;
+  final OnListAdd? onListAdd;
 
   /// Set in order to provide custom acceptance criteria for when a list can be
   /// dropped onto a specific other list
-  final ListOnWillAccept listOnWillAccept;
+  final ListOnWillAccept? listOnWillAccept;
 
   /// Set in order to get the lists involved in a drag and drop operation after
   /// a list has been accepted. For general use cases where only reordering is
   /// necessary, only [onListReorder] or [onListAdd] is needed, and this should
   /// be left null. [onListReorder] or [onListAdd] will be called after this.
-  final ListOnAccept listOnAccept;
+  final ListOnAccept? listOnAccept;
 
   /// Set in order to provide custom acceptance criteria for when a list can be
   /// dropped onto a specific target. This target always exists as the last
   /// target the DragAndDropLists, and also can be used independently.
-  final ListTargetOnWillAccept listTargetOnWillAccept;
+  final ListTargetOnWillAccept? listTargetOnWillAccept;
 
   /// Set in order to get the list and target involved in a drag and drop
   /// operation after a list has been accepted. For general use cases where only
   /// reordering is necessary, only [onListReorder] or [onListAdd] is needed,
   /// and this should be left null. [onListReorder] or [onListAdd] will be
   /// called after this.
-  final ListTargetOnAccept listTargetOnAccept;
+  final ListTargetOnAccept? listTargetOnAccept;
 
   /// Called when a list dragging is starting or ending
-  final OnListDraggingChanged onListDraggingChanged;
+  final OnListDraggingChanged? onListDraggingChanged;
 
   /// Set in order to provide custom acceptance criteria for when a item can be
   /// dropped onto a specific other item
-  final ItemOnWillAccept itemOnWillAccept;
+  final ItemOnWillAccept? itemOnWillAccept;
 
   /// Set in order to get the items involved in a drag and drop operation after
   /// an item has been accepted. For general use cases where only reordering is
   /// necessary, only [onItemReorder] or [onItemAdd] is needed, and this should
   /// be left null. [onItemReorder] or [onItemAdd] will be called after this.
-  final ItemOnAccept itemOnAccept;
+  final ItemOnAccept? itemOnAccept;
 
   /// Set in order to provide custom acceptance criteria for when a item can be
   /// dropped onto a specific target. This target always exists as the last
   /// target for list of items, and also can be used independently.
-  final ItemTargetOnWillAccept itemTargetOnWillAccept;
+  final ItemTargetOnWillAccept? itemTargetOnWillAccept;
 
   /// Set in order to get the item and target involved in a drag and drop
   /// operation after a item has been accepted. For general use cases where only
   /// reordering is necessary, only [onItemReorder] or [onItemAdd] is needed,
   /// and this should be left null. [onItemReorder] or [onItemAdd] will be
   /// called after this.
-  final ItemTargetOnAccept itemTargetOnAccept;
+  final ItemTargetOnAccept? itemTargetOnAccept;
 
   /// Called when an item dragging is starting or ending
-  final OnItemDraggingChanged onItemDraggingChanged;
+  final OnItemDraggingChanged? onItemDraggingChanged;
 
   /// Width of a list item when it is being dragged.
-  final double itemDraggingWidth;
+  final double? itemDraggingWidth;
 
   /// The widget that will be displayed at a potential drop position in a list
   /// when an item is being dragged.
-  final Widget itemGhost;
+  final Widget? itemGhost;
 
   /// The opacity of the [itemGhost]. This must be between 0 and 1.
   final double itemGhostOpacity;
@@ -171,22 +171,22 @@ class DragAndDropLists extends StatefulWidget {
   final bool itemDragOnLongPress;
 
   /// The decoration surrounding an item while it is in the process of being dragged.
-  final Decoration itemDecorationWhileDragging;
+  final Decoration? itemDecorationWhileDragging;
 
   /// A widget that will be displayed between each individual item.
-  final Widget itemDivider;
+  final Widget? itemDivider;
 
   /// The width of a list when dragging.
-  final double listDraggingWidth;
+  final double? listDraggingWidth;
 
   /// The widget to be displayed as the last element in the DragAndDropLists,
   /// where a list will be accepted as the last list.
-  final Widget listTarget;
+  final Widget? listTarget;
 
   /// The widget to be displayed at a potential list position while a list is being dragged.
   /// This must not be null when [children] includes one or more
   /// [DragAndDropListExpansion] or other class that inherit from [DragAndDropListExpansionInterface].
-  final Widget listGhost;
+  final Widget? listGhost;
 
   /// The opacity of [listGhost]. It must be between 0 and 1.
   final double listGhostOpacity;
@@ -201,25 +201,25 @@ class DragAndDropLists extends StatefulWidget {
   final bool listDragOnLongPress;
 
   /// The decoration surrounding a list.
-  final Decoration listDecoration;
+  final Decoration? listDecoration;
 
   /// The decoration surrounding a list while it is in the process of being dragged.
-  final Decoration listDecorationWhileDragging;
+  final Decoration? listDecorationWhileDragging;
 
   /// The decoration surrounding the inner list of items.
-  final Decoration listInnerDecoration;
+  final Decoration? listInnerDecoration;
 
   /// A widget that will be displayed between each individual list.
-  final Widget listDivider;
+  final Widget? listDivider;
 
   /// Whether it should put a divider on the last list or not.
   final bool listDividerOnLastChild;
 
   /// The padding between each individual list.
-  final EdgeInsets listPadding;
+  final EdgeInsets? listPadding;
 
   /// A widget that will be displayed whenever a list contains no items.
-  final Widget contentsWhenEmpty;
+  final Widget? contentsWhenEmpty;
 
   /// The width of each individual list. This must be set to a finite value when
   /// [axis] is set to Axis.horizontal.
@@ -258,7 +258,7 @@ class DragAndDropLists extends StatefulWidget {
 
   /// A scroll controller that can be used for the scrolling of the first level lists.
   /// This must be set if [sliverList] is set to true.
-  final ScrollController scrollController;
+  final ScrollController? scrollController;
 
   /// Set to true in order to disable all scrolling of the lists.
   /// Note: to disable scrolling for sliver lists, it is also necessary in your
@@ -267,7 +267,7 @@ class DragAndDropLists extends StatefulWidget {
 
   /// Set a custom drag handle to use iOS-like handles to drag rather than long
   /// or short presses
-  final Widget dragHandle;
+  final Widget? dragHandle;
 
   /// Set the drag handle to be on the left side instead of the default right side
   final bool dragHandleOnLeft;
@@ -334,10 +334,10 @@ class DragAndDropLists extends StatefulWidget {
     this.listDragHandleVerticalAlignment = DragHandleVerticalAlignment.top,
     this.itemDragHandleVerticalAlignment = DragHandleVerticalAlignment.center,
     this.constrainDraggingAxis = true,
-    Key key,
+    Key? key,
   }) : super(key: key) {
     if (listGhost == null &&
-        children
+        children!
             .where((element) => element is DragAndDropListExpansionInterface)
             .isNotEmpty)
       throw Exception(
@@ -361,10 +361,10 @@ class DragAndDropLists extends StatefulWidget {
 }
 
 class DragAndDropListsState extends State<DragAndDropLists> {
-  ScrollController _scrollController;
+  ScrollController? _scrollController;
   bool _pointerDown = false;
-  double _pointerYPosition;
-  double _pointerXPosition;
+  double? _pointerYPosition;
+  double? _pointerXPosition;
   bool _scrolling = false;
   PageStorageBucket _pageStorageBucket = PageStorageBucket();
 
@@ -428,7 +428,7 @@ class DragAndDropListsState extends State<DragAndDropLists> {
       lastListTargetSize: widget.lastListTargetSize,
     );
 
-    if (widget.children != null && widget.children.isNotEmpty) {
+    if (widget.children != null && widget.children!.isNotEmpty) {
       Widget outerListHolder;
 
       if (widget.sliverList) {
@@ -440,7 +440,7 @@ class DragAndDropListsState extends State<DragAndDropLists> {
         outerListHolder = _buildListView(parameters, dragAndDropListTarget);
       }
 
-      if (widget.children
+      if (widget.children!
           .where((e) => e is DragAndDropListExpansionInterface)
           .isNotEmpty) {
         outerListHolder = PageStorage(
@@ -482,11 +482,11 @@ class DragAndDropListsState extends State<DragAndDropLists> {
       DragAndDropBuilderParameters parameters) {
     if (widget.axis == Axis.vertical) {
       return Column(
-        children: _buildOuterList(dragAndDropListTarget, parameters),
+        children: _buildOuterList(dragAndDropListTarget, parameters) as List<Widget>,
       );
     } else {
       return Row(
-        children: _buildOuterList(dragAndDropListTarget, parameters),
+        children: _buildOuterList(dragAndDropListTarget, parameters) as List<Widget>,
       );
     }
   }
@@ -496,11 +496,11 @@ class DragAndDropListsState extends State<DragAndDropLists> {
     return ListView(
       scrollDirection: widget.axis,
       controller: _scrollController,
-      children: _buildOuterList(dragAndDropListTarget, parameters),
+      children: _buildOuterList(dragAndDropListTarget, parameters) as List<Widget>,
     );
   }
 
-  List<Widget> _buildOuterList(DragAndDropListTarget dragAndDropListTarget,
+  List<Widget?> _buildOuterList(DragAndDropListTarget dragAndDropListTarget,
       DragAndDropBuilderParameters parameters) {
     bool includeSeparators = widget.listDivider != null;
     int childrenCount = _calculateChildrenCount(includeSeparators);
@@ -520,7 +520,7 @@ class DragAndDropListsState extends State<DragAndDropLists> {
       return (widget.children?.length ?? 0) + 1;
   }
 
-  Widget _buildInnerList(
+  Widget? _buildInnerList(
       int index,
       int childrenCount,
       DragAndDropListTarget dragAndDropListTarget,
@@ -533,7 +533,7 @@ class DragAndDropListsState extends State<DragAndDropLists> {
     } else {
       return DragAndDropListWrapper(
         dragAndDropList:
-            widget.children[(includeSeparators ? index / 2 : index).toInt()],
+            widget.children![(includeSeparators ? index / 2 : index).toInt()],
         parameters: parameters,
       );
     }
@@ -541,7 +541,7 @@ class DragAndDropListsState extends State<DragAndDropLists> {
 
   _internalOnItemReorder(DragAndDropItem reordered, DragAndDropItem receiver) {
     if (widget.itemOnAccept != null) {
-      widget.itemOnAccept(reordered, receiver);
+      widget.itemOnAccept!(reordered, receiver);
     }
 
     int reorderedListIndex = -1;
@@ -549,15 +549,15 @@ class DragAndDropListsState extends State<DragAndDropLists> {
     int receiverListIndex = -1;
     int receiverItemIndex = -1;
 
-    for (int i = 0; i < widget.children.length; i++) {
+    for (int i = 0; i < widget.children!.length; i++) {
       if (reorderedItemIndex == -1) {
         reorderedItemIndex =
-            widget.children[i].children.indexWhere((e) => reordered == e);
+            widget.children![i].children!.indexWhere((e) => reordered == e);
         if (reorderedItemIndex != -1) reorderedListIndex = i;
       }
       if (receiverItemIndex == -1) {
         receiverItemIndex =
-            widget.children[i].children.indexWhere((e) => receiver == e);
+            widget.children![i].children!.indexWhere((e) => receiver == e);
         if (receiverItemIndex != -1) receiverListIndex = i;
       }
       if (reorderedItemIndex != -1 && receiverItemIndex != -1) {
@@ -568,7 +568,7 @@ class DragAndDropListsState extends State<DragAndDropLists> {
     if (reorderedItemIndex == -1) {
       // this is a new item
       if (widget.onItemAdd != null)
-        widget.onItemAdd(reordered, receiverListIndex, receiverItemIndex);
+        widget.onItemAdd!(reordered, receiverListIndex, receiverItemIndex);
     } else {
       if (reorderedListIndex == receiverListIndex &&
           receiverItemIndex > reorderedItemIndex) {
@@ -577,37 +577,37 @@ class DragAndDropListsState extends State<DragAndDropLists> {
       }
 
       if (widget.onItemReorder != null)
-        widget.onItemReorder(reorderedItemIndex, reorderedListIndex,
+        widget.onItemReorder!(reorderedItemIndex, reorderedListIndex,
             receiverItemIndex, receiverListIndex);
     }
   }
 
   _internalOnListReorder(
-      DragAndDropListInterface reordered, DragAndDropListInterface receiver) {
-    int reorderedListIndex = widget.children.indexWhere((e) => reordered == e);
-    int receiverListIndex = widget.children.indexWhere((e) => receiver == e);
+      DragAndDropListInterface reordered, DragAndDropListInterface? receiver) {
+    int reorderedListIndex = widget.children!.indexWhere((e) => reordered == e);
+    int receiverListIndex = widget.children!.indexWhere((e) => receiver == e);
 
     int newListIndex = receiverListIndex;
 
-    if (widget.listOnAccept != null) widget.listOnAccept(reordered, receiver);
+    if (widget.listOnAccept != null) widget.listOnAccept!(reordered, receiver!);
 
     if (reorderedListIndex == -1) {
       // this is a new list
-      if (widget.onListAdd != null) widget.onListAdd(reordered, newListIndex);
+      if (widget.onListAdd != null) widget.onListAdd!(reordered, newListIndex);
     } else {
       if (newListIndex > reorderedListIndex) {
         // same list, so if the new position is after the old position, the removal of the old item must be taken into account
         newListIndex--;
       }
       if (widget.onListReorder != null)
-        widget.onListReorder(reorderedListIndex, newListIndex);
+        widget.onListReorder!(reorderedListIndex, newListIndex);
     }
   }
 
   _internalOnItemDropOnLastTarget(DragAndDropItem newOrReordered,
-      DragAndDropListInterface parentList, DragAndDropItemTarget receiver) {
+      DragAndDropListInterface? parentList, DragAndDropItemTarget? receiver) {
     if (widget.itemTargetOnAccept != null) {
-      widget.itemTargetOnAccept(newOrReordered, parentList, receiver);
+      widget.itemTargetOnAccept!(newOrReordered, parentList!, receiver!);
     }
 
     int reorderedListIndex = -1;
@@ -615,18 +615,18 @@ class DragAndDropListsState extends State<DragAndDropLists> {
     int receiverListIndex = -1;
     int receiverItemIndex = -1;
 
-    if (widget.children != null && widget.children.isNotEmpty) {
-      for (int i = 0; i < widget.children.length; i++) {
+    if (widget.children != null && widget.children!.isNotEmpty) {
+      for (int i = 0; i < widget.children!.length; i++) {
         if (reorderedItemIndex == -1) {
-          reorderedItemIndex = widget.children[i].children
+          reorderedItemIndex = widget.children![i].children
                   ?.indexWhere((e) => newOrReordered == e) ??
               -1;
           if (reorderedItemIndex != -1) reorderedListIndex = i;
         }
 
-        if (receiverItemIndex == -1 && widget.children[i] == parentList) {
+        if (receiverItemIndex == -1 && widget.children![i] == parentList) {
           receiverListIndex = i;
-          receiverItemIndex = widget.children[i].children?.length ?? -1;
+          receiverItemIndex = widget.children![i].children?.length ?? -1;
         }
 
         if (reorderedItemIndex != -1 && receiverItemIndex != -1) {
@@ -637,7 +637,7 @@ class DragAndDropListsState extends State<DragAndDropLists> {
 
     if (reorderedItemIndex == -1) {
       if (widget.onItemAdd != null)
-        widget.onItemAdd(newOrReordered, receiverListIndex, reorderedItemIndex);
+        widget.onItemAdd!(newOrReordered, receiverListIndex, reorderedItemIndex);
     } else {
       if (reorderedListIndex == receiverListIndex &&
           receiverItemIndex > reorderedItemIndex) {
@@ -645,7 +645,7 @@ class DragAndDropListsState extends State<DragAndDropLists> {
         receiverItemIndex--;
       }
       if (widget.onItemReorder != null)
-        widget.onItemReorder(reorderedItemIndex, reorderedListIndex,
+        widget.onItemReorder!(reorderedItemIndex, reorderedListIndex,
             receiverItemIndex, receiverListIndex);
     }
   }
@@ -654,17 +654,17 @@ class DragAndDropListsState extends State<DragAndDropLists> {
       DragAndDropListInterface newOrReordered, DragAndDropListTarget receiver) {
     // determine if newOrReordered is new or existing
     int reorderedListIndex =
-        widget.children.indexWhere((e) => newOrReordered == e);
+        widget.children!.indexWhere((e) => newOrReordered == e);
 
     if (widget.listOnAccept != null)
-      widget.listTargetOnAccept(newOrReordered, receiver);
+      widget.listTargetOnAccept!(newOrReordered, receiver);
 
     if (reorderedListIndex >= 0) {
       if (widget.onListReorder != null)
-        widget.onListReorder(reorderedListIndex, widget.children.length - 1);
+        widget.onListReorder!(reorderedListIndex, widget.children!.length - 1);
     } else {
       if (widget.onListAdd != null)
-        widget.onListAdd(newOrReordered, reorderedListIndex);
+        widget.onListAdd!(newOrReordered, reorderedListIndex);
     }
   }
 
@@ -698,10 +698,10 @@ class DragAndDropListsState extends State<DragAndDropLists> {
       double step = 1.5;
       double overDragMax = 20.0;
       double overDragCoefficient = 5.0;
-      double newOffset;
+      double? newOffset;
 
-      var rb = context.findRenderObject();
-      Size size;
+      var rb = context.findRenderObject()!;
+      late Size size;
       if (rb is RenderBox)
         size = rb.size;
       else if (rb is RenderSliver) size = rb.paintBounds.size;
@@ -712,53 +712,53 @@ class DragAndDropListsState extends State<DragAndDropLists> {
         double top = topLeftOffset.dy;
         double bottom = bottomRightOffset.dy;
 
-        if (_pointerYPosition < (top + scrollAreaSize) &&
-            _scrollController.position.pixels >
-                _scrollController.position.minScrollExtent) {
+        if (_pointerYPosition! < (top + scrollAreaSize) &&
+            _scrollController!.position.pixels >
+                _scrollController!.position.minScrollExtent) {
           final overDrag =
-              max((top + scrollAreaSize) - _pointerYPosition, overDragMax);
+              max((top + scrollAreaSize) - _pointerYPosition!, overDragMax);
           newOffset = max(
-              _scrollController.position.minScrollExtent,
-              _scrollController.position.pixels -
+              _scrollController!.position.minScrollExtent,
+              _scrollController!.position.pixels -
                   step * overDrag / overDragCoefficient);
-        } else if (_pointerYPosition > (bottom - scrollAreaSize) &&
-            _scrollController.position.pixels <
-                _scrollController.position.maxScrollExtent) {
+        } else if (_pointerYPosition! > (bottom - scrollAreaSize) &&
+            _scrollController!.position.pixels <
+                _scrollController!.position.maxScrollExtent) {
           final overDrag = max<double>(
-              _pointerYPosition - (bottom - scrollAreaSize), overDragMax);
+              _pointerYPosition! - (bottom - scrollAreaSize), overDragMax);
           newOffset = min(
-              _scrollController.position.maxScrollExtent,
-              _scrollController.position.pixels +
+              _scrollController!.position.maxScrollExtent,
+              _scrollController!.position.pixels +
                   step * overDrag / overDragCoefficient);
         }
       } else {
         double left = topLeftOffset.dx;
         double right = bottomRightOffset.dx;
 
-        if (_pointerXPosition < (left + scrollAreaSize) &&
-            _scrollController.position.pixels >
-                _scrollController.position.minScrollExtent) {
+        if (_pointerXPosition! < (left + scrollAreaSize) &&
+            _scrollController!.position.pixels >
+                _scrollController!.position.minScrollExtent) {
           final overDrag =
-              max((left + scrollAreaSize) - _pointerXPosition, overDragMax);
+              max((left + scrollAreaSize) - _pointerXPosition!, overDragMax);
           newOffset = max(
-              _scrollController.position.minScrollExtent,
-              _scrollController.position.pixels -
+              _scrollController!.position.minScrollExtent,
+              _scrollController!.position.pixels -
                   step * overDrag / overDragCoefficient);
-        } else if (_pointerXPosition > (right - scrollAreaSize) &&
-            _scrollController.position.pixels <
-                _scrollController.position.maxScrollExtent) {
+        } else if (_pointerXPosition! > (right - scrollAreaSize) &&
+            _scrollController!.position.pixels <
+                _scrollController!.position.maxScrollExtent) {
           final overDrag = max<double>(
-              _pointerYPosition - (right - scrollAreaSize), overDragMax);
+              _pointerYPosition! - (right - scrollAreaSize), overDragMax);
           newOffset = min(
-              _scrollController.position.maxScrollExtent,
-              _scrollController.position.pixels +
+              _scrollController!.position.maxScrollExtent,
+              _scrollController!.position.pixels +
                   step * overDrag / overDragCoefficient);
         }
       }
 
       if (newOffset != null) {
         _scrolling = true;
-        await _scrollController.animateTo(newOffset,
+        await _scrollController!.animateTo(newOffset,
             duration: Duration(milliseconds: duration), curve: Curves.linear);
         _scrolling = false;
         if (_pointerDown) _scrollList();
@@ -767,7 +767,7 @@ class DragAndDropListsState extends State<DragAndDropLists> {
   }
 
   static Offset localToGlobal(RenderObject object, Offset point,
-      {RenderObject ancestor}) {
+      {RenderObject? ancestor}) {
     return MatrixUtils.transformPoint(object.getTransformTo(ancestor), point);
   }
 }
