@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 
-typedef void OnWidgetSizeChange(Size size);
+typedef void OnWidgetSizeChange(Size? size);
 
 class MeasureSize extends StatefulWidget {
-  final Widget child;
+  final Widget? child;
   final OnWidgetSizeChange onSizeChange;
 
   const MeasureSize({
-    Key key,
-    @required this.onSizeChange,
-    @required this.child,
+    Key? key,
+    required this.onSizeChange,
+    required this.child,
   }) : super(key: key);
 
   @override
@@ -20,7 +20,7 @@ class MeasureSize extends StatefulWidget {
 class _MeasureSizeState extends State<MeasureSize> {
   @override
   Widget build(BuildContext context) {
-    SchedulerBinding.instance.addPostFrameCallback(postFrameCallback);
+    SchedulerBinding.instance!.addPostFrameCallback(postFrameCallback);
     return Container(
       key: widgetKey,
       child: widget.child,
